Adding NFLdb sql files
Thank you to Jason and his friend for pointing us to another great Github project and providing some scripts for us to work with it. 

https://github.com/BurntSushi/nfldb  <--- This repo should have scripts and info for us to build our own local databases on our personal machines if we choose to (still exploring options with shared db access from aws). I think it would populate a database with roughly 10Gb of data.

I haven't actually tried to use this stuff yet, so the following is inference and conjecture:
The nfl_SQL_code.sql file ports the data from Postgres database to an SQL database. If you prefer Postgres, no need to use this file, but you will probably have to change the next couple of files up for them to work. This is all assuming that you've used the code in the github repo to make your Postgres Db. 

game_data_view.sql and player_only_stats_view.sql are scripts that should nicely display some data from the database you just created. The language isn't pure SQL, it's